be fix

+++ server/api/token/paseto_token.go
@@ -19,7 +19,7 @@ func NewPasetoMaker(symmetricKey string) (Maker, error) {
 }
 
 // CreateToken creates a new token
-func (maker *PasetoMaker) CreateToken(userID int64, accountType int64, duration time.Duration) (string, error) {
+func (maker *PasetoMaker) CreateToken(userID int32, accountType int32, duration time.Duration) (string, error) {
 	payload, err := NewPayload(userID, accountType, duration)
 	if err != nil {
 		return "", err
